% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_subgraphs}
\alias{bind_subgraphs}
\title{Bind Subgraphs into a Single Graph}
\usage{
bind_subgraphs(subgraphs)
}
\arguments{
\item{subgraphs}{A list of subgraphs.}
}
\value{
A single graph created by binding the subgraphs together.
}
\description{
This function takes a list of subgraphs and binds them together into a single graph.
It performs the following steps:
\itemize{
\item Converts each subgraph into a data frame representation.
\item Binds the data frames together using \code{bind_rows()}.
\item Constructs a new graph from the combined data frame using \code{graph_from_data_frame()}.
}
}
