% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bridge_remover}
\alias{bridge_remover}
\title{Remove Bridges and Isolated Vertices from a Graph}
\usage{
bridge_remover(graph)
}
\arguments{
\item{graph}{The input graph.}
}
\value{
The pruned graph without bridges and isolated vertices.
}
\description{
This function removes bridges (edges whose removal would increase the number
of connected components) and isolated vertices (vertices with no incident edges)
from a graph.
}
