% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{calculate_jaccard_index}
\alias{calculate_jaccard_index}
\title{Calculate Jaccard Index for Pairs}
\usage{
calculate_jaccard_index(tbl, ontology, gene_set)
}
\arguments{
\item{tbl}{Tibble containing the gene sets.}

\item{ontology}{Name of the ontology column in the tibble}

\item{gene_set}{Name of the gene set column in the tibble}
}
\value{
A tibble with pairs of gene sets, associated ontologies, and Jaccard index values.
}
\description{
Calculates the Jaccard index for pairs of gene sets based on the provided ontology and gene set.
}
