% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corshrink_word2vec.R
\name{corshrink_word2vec}
\alias{corshrink_word2vec}
\title{Adaptive Shrinkage of cosine similarities in word2vec output.}
\usage{
corshrink_word2vec(model_true, model_boot_list, word_vec,
  num_related_words = 500)
}
\arguments{
\item{model_true}{A word2vec VectorSpace model output obtained by fitting the model on corpus.}

\item{model_boot_list}{A list of VectorSpace models obtained by fitting the word2vec on resampled
corpus data.}

\item{word_vec}{A word set or a vector of words of interest. Shrinkage of cosine similarities
for the primary linked words to each word in this word set.}

\item{num_related_words}{The number of primary linked words to each word taken into the model.}
}
\value{
Returns a list with each element corresponding to ash results for each word
       in \code{word_vec}. An element of the list is anopther list containing the following
       features.
            \code{similar_words} : all the words taken into the shrinkage model for
                                   each word in \code{word_vec}.
             \code{cosine_est}: cosine similarities from the original word2vec model
             \code{ sd_cosine_transform_est}: standard error of Fisher z-scores obtained from
                                      cosine similarities of primary links with each word
                                      in \code{word_vec} from resampled corpus data.
             \code{ash_out}: The ash result for each word in \code{word_vec}
              \code{ash_cosine_est}: ash shrunk cosine similarities for each word in
                                     \code{word_vec}.
}
\description{
Performs an adaptive shrinkage (ash) of the cosine similarities in word2vec output.
             The ash framework has been proposed by Matthew Stephens (2016).
}
