% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_graph}
\alias{generate_graph}
\title{Generate Graph}
\usage{
generate_graph(tbl, n = NULL, prop = 1)
}
\arguments{
\item{tbl}{Tibble containing the graph information.}

\item{n}{Maximum number of rows to include in the graph.}

\item{prop}{Proportion of rows to include in the graph.}
}
\value{
A graph object representing the generated graph.
}
\description{
Generates a graph from a tibble based on the specified criteria.
}
