% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{generate_pairs}
\alias{generate_pairs}
\title{Generate Pairs}
\usage{
generate_pairs(tbl, ontology, gene_set)
}
\arguments{
\item{tbl}{Tibble containing the gene sets.}

\item{ontology}{Name of the ontology column in the tibble}

\item{gene_set}{Name of the gene set column in the tibble}
}
\value{
A tibble with pairs of gene sets and associated ontologies.
}
\description{
Generates pairs of gene sets from a tibble based on the provided ontology and gene set.
}
