\name{internal_use_only}

\alias{CACHE_DIR}
\alias{cache_expr}
\alias{CACHE_EXT}
\alias{deps_changed}
\alias{eval_and_cache}
\alias{expr_printer}
\alias{findDeps}
\alias{get_cache_dir}
\alias{get_chunk_id}
\alias{get_expr_hash}
\alias{load_from_cache_env}
\alias{log_debug}
\alias{make_cache_expr}
\alias{resetStorage}
\alias{weaverEvalWithOpt}
\alias{weaverLatexFinish}
\alias{weaver_opts}
\alias{weaverRemoveOrphans}
\alias{weaverRuncode}

\title{Internal functions for the weaver package}

\description{
 These functions are not part of the public API of the package.
}

\keyword{utilities}
