% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jaccard_index}
\alias{jaccard_index}
\title{Calculate Jaccard Index}
\usage{
jaccard_index(set1, set2)
}
\arguments{
\item{set1}{First set.}

\item{set2}{Second set.}
}
\value{
Jaccard index value.
}
\description{
Calculates the Jaccard index between two sets.
}
