% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankers.R
\name{normalized_degree_ranker}
\alias{normalized_degree_ranker}
\title{Calculate Normalized Degree Rank for Graph Ontologies}
\usage{
normalized_degree_ranker(graph)
}
\arguments{
\item{graph}{The input graph.}
}
\value{
A tibble containing the ontologies and their normalized degree ranks.
}
\description{
This function calculates the normalized degree rank for graph ontologies.
It takes a graph as input and performs the following steps:
\itemize{
\item Calculates the degree centrality for the graph, considering both incoming and outgoing edges.
\item Normalizes the degree centrality values.
\item Returns the ontologies along with their corresponding normalized degree ranks in descending order.
}
}
