% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankers.R
\name{normalized_jaccard_ranker}
\alias{normalized_jaccard_ranker}
\title{Calculate Normalized Jaccard Rank for Subgraph Ontologies}
\usage{
normalized_jaccard_ranker(subgraph)
}
\arguments{
\item{subgraph}{The input subgraph.}
}
\value{
A data frame containing the ranked ontologies and their Jaccard ranks.
}
\description{
This function calculates the normalized Jaccard rank for subgraph ontologies.
It takes a subgraph as input and performs the following steps:
\itemize{
\item Converts the subgraph to a long-format data frame.
\item Counts the number of edges for each ontology in the subgraph.
\item Ranks the ontologies based on the Jaccard similarity index.
\item Returns the ranked ontologies in descending order of Jaccard rank.
}
}
