% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot Graph}
\usage{
plot_graph(graph, remove_bridges = FALSE, min_vertices = 3)
}
\arguments{
\item{graph}{The input graph.}

\item{remove_bridges}{Logical indicating whether to remove bridges from the graph (default is FALSE).}

\item{min_vertices}{The minimum number of vertices required for subgraph decomposition (default is 3).}
}
\description{
This function plots a graph using the ggraph package.
It performs the following steps:
\itemize{
\item Decomposes the input graph into subgraphs based on the specified minimum number of vertices using \code{decompose()}.
\item Binds the subgraphs together into a single graph using the \code{bind_subgraphs()} function.
\item Optionally removes bridges from the graph using the \code{bridge_remover()} function.
\item Plots the resulting graph using the Fruchterman-Reingold layout in ggraph.
\item Uses alpha blending of edges based on the index attribute.
\item Colors nodes as light grey with a larger size for the subgraph nodes, and black with a smaller size for other nodes.
\item Applies a void theme to the plot and removes the legend.
}
}
