% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankers.R
\name{rank_subgraph_nodes}
\alias{rank_subgraph_nodes}
\title{Rank Nodes in Subgraphs of a Graph}
\usage{
rank_subgraph_nodes(graph, ranker, min_vertices = 3, remove_bridges = FALSE)
}
\arguments{
\item{graph}{The input graph.}

\item{ranker}{A ranking function that takes a subgraph as input and returns ranked nodes.}

\item{min_vertices}{The minimum number of vertices required to form a subgraph (default: 3).}

\item{remove_bridges}{Logical indicating whether to remove bridges from each subgraph (default: FALSE).}
}
\value{
A tibble containing the ranked nodes and their corresponding subgraph indices.
}
\description{
This function ranks the nodes in subgraphs of a graph based on a specified ranking function.
It takes a graph, a ranking function, and optional parameters as input and performs the following steps:
\itemize{
\item Decomposes the graph into subgraphs based on a minimum number of vertices.
\item If specified, removes bridges from each subgraph.
\item Applies the ranking function to each subgraph and retrieves the ranked nodes.
\item Returns a tibble containing the ranked nodes along with the corresponding subgraph indices.
}
}
