% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankers.R
\name{scaled_eigenvector_ranker}
\alias{scaled_eigenvector_ranker}
\title{Calculate Scaled Eigenvector Rank for Graph Ontologies}
\usage{
scaled_eigenvector_ranker(graph)
}
\arguments{
\item{graph}{The input graph.}
}
\value{
A tibble containing the ontologies and their scaled eigenvector ranks.
}
\description{
This function calculates the scaled eigenvector rank for graph ontologies.
It takes a graph as input and performs the following steps:
\itemize{
\item Calculates the eigenvector centrality for the graph.
\item Scales the eigenvector centrality values.
\item Returns the ontologies along with their corresponding eigenvector ranks
in descending order.
}
}
