% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsea.R
\name{wsea}
\alias{wsea}
\title{Word Set Enrichment Analysis (WSEA) across corpora}
\usage{
wsea(model_list, word_sets, header_set, model_names = NULL,
  add_groups = NULL, numwords = 1000, fgsea.control = list())
}
\arguments{
\item{model_list}{A list of VectorSpace models for a number of corpora. The length of the
list is equal to the number of corpora.}

\item{word_sets}{A list of vector of words (word sets) whose enrichment is to be studied.}

\item{header_set}{a word or a vector of words defining a topic of interest.}

\item{model_names}{The names of the corpora that are compared. Defaults to numerical groups.}

\item{add_groups}{A list of vectors indicating group labels that the user wants to compare.
Using \code{add_groups}, the user can compare a set of corpora with another
set of corpora beyod the 1v1 and 1vall comparisons the default version does.}

\item{numwords}{The number of neighboring words to each word in the word sets pooled into the
word set enrichment analysis.}

\item{fgsea.control}{The control parameters for the GSEA model fitting using the fgsea() package.}
}
\value{
Produces 1v1, 1vall WSEA output for all copora (each output similar to the fgsea output)
        together with group based WSEA output as specified by the \code{add_groups}.
}
\description{
Performs a word set enrichment across multiple corpora in relation to a
             topic of interest, specified by the \code{header_set}.
}
