% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wssa.R
\name{wssa}
\alias{wssa}
\title{Word Set Significane Analysis (WSSA) for a given word set in a corpus.}
\usage{
wssa(model, model_boot = NULL, word_set, header_set, control_set = NULL,
  num_controls = 5000, fgsea.control = list())
}
\arguments{
\item{model}{A word2vec VectorSpace model ouput.}

\item{model_boot}{If not NULL, a list of resampled Vectospace model outputs generated by
applying word2vec on resampled corpus.Defaults to NULL, which ignores the
resampling.}

\item{word_set}{A set of words (a vector of words) whose significance is to be tested.}

\item{header_set}{a word or a vector of words defining a topic of interest.}

\item{control_set}{A vector of words forming the background. Defaults to NULL in which case
the function chooses a random set of words from the vocabulary.}

\item{num_controls}{The number of control words chosen for background if \code{control_set}
is NULL.}

\item{fgsea.control}{The control parameters for the GSEA model fitting using the fgsea() package.}
}
\value{
Returns an output of the significance analysis (p-values and expected score) for the
        \code{word_set} of interest in the context of the \code{header_set} with the
        \code{control_set} as background.
}
\description{
Performs a word set significance analysis in a corpus for a given
             topic of interest, specified by the \code{header_set}, with a \code{control_set}
             forming the background.
}
